Fix set_motd bugs

+++ .files/motd/set_motd.sh
@@ -1,14 +1,18 @@
-#!/usr/env/bin bash
+#!/usr/bin/env bash
 
-# Run this script directly from its location.
-test -f ./motd.sh
+if [ ! -f ./motd.sh ]; then
+    echo 'Run set_motd.sh directly from its location.'
+    exit 1
+fi
 
-test -f /etc/motd
-sudo truncate -s 0 /etc/motd
+if [ ! -f /etc/motd ]; then
+    echo 'No message of the day detected. Setup "/etc/motd" manually.'
+    exit 1
+fi
 
+sudo truncate -s 0 /etc/motd || exit 1
 motd=/etc/update-motd.d/10-motd
 test -f "$motd" || sudo rm "$motd"
-
 cp ./motd.sh "$motd"
 chmod +x "$motd"
